Move pkg/config -> pkg/processor/config
