+++ internal/pkg/processor/processor.go
@@ -5,10 +5,10 @@ import (
 
 	"github.com/dustin/go-humanize"
 
-	"github.com/keboola/processor-split-table/internal/pkg/config"
 	"github.com/keboola/processor-split-table/internal/pkg/finder"
 	"github.com/keboola/processor-split-table/internal/pkg/kbc"
 	"github.com/keboola/processor-split-table/internal/pkg/log"
+	"github.com/keboola/processor-split-table/internal/pkg/processor/config"
 	"github.com/keboola/processor-split-table/internal/pkg/slicer"
 	"github.com/keboola/processor-split-table/internal/pkg/utils"
 )
